Refactor Simulation_robin and unify experiment outputs
